+++ Tournament-Manager/TournamentManagement/NHL/TeamNHL.cs
@@ -9,7 +9,7 @@ namespace TournamentManagement.NHL
 {
     public class TeamNHL : Team
     {
-        public TeamNHL(string name, string shortName, float winRate) : base(name, shortName, winRate)
+        public TeamNHL(string name, string shortName, double winRate) : base(name, shortName, winRate)
         {
                 
         }
